+++ tests/corpus/validation/test_base.py
@@ -1,10 +1,10 @@
 from audiomate.corpus import validation
 
 
-class TestInvalidUtterancesResult:
+class TestInvalidItemsResult:
 
     def test_get_report_on_pass(self):
-        result = validation.InvalidUtterancesResult(passed=True, invalid_utterances={}, name='MyName', info={'a': 'b'})
+        result = validation.InvalidItemsResult(passed=True, invalid_items={}, name='MyName', info={'a': 'b'})
 
         assert result.get_report() == '\n'.join([
             'MyName',
@@ -16,7 +16,7 @@ class TestInvalidUtterancesResult:
         ])
 
     def test_get_report_on_fail(self):
-        result = validation.InvalidUtterancesResult(passed=False, invalid_utterances={
+        result = validation.InvalidItemsResult(passed=False, invalid_items={
             'utt-1': 'reason1',
             'utt-4': 'reason4',
             'utt-2': 3
@@ -31,7 +31,35 @@ class TestInvalidUtterancesResult:
             '',
             'Result: Failed',
             '',
-            'Invalid utterances:',
+            'Invalid Utterances:',
+            '    * utt-1 (reason1)',
+            '    * utt-2 (3)',
+            '    * utt-4 (reason4)'
+        ])
+
+    def test_get_report_on_fail_with_type_name(self):
+        result = validation.InvalidItemsResult(
+            passed=False,
+            invalid_items={
+                'utt-1': 'reason1',
+                'utt-4': 'reason4',
+                'utt-2': 3
+            },
+            name='MyName',
+            item_name='Tracks',
+            info={'c': 'b', 'a': 49.0}
+        )
+
+        assert result.get_report() == '\n'.join([
+            'MyName',
+            '======',
+            '',
+            '--> a: 49.0',
+            '--> c: b',
+            '',
+            'Result: Failed',
+            '',
+            'Invalid Tracks:',
             '    * utt-1 (reason1)',
             '    * utt-2 (3)',
             '    * utt-4 (reason4)'